pyproject.toml
[tool.poetry]
name = "micromesh"
version = "0.1.0"
description = ""
authors = ["warhammer21 <shreyakvashisht21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0"
psycopg2-binary = "^2.9"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
